+++ src/pages/matching/index.tsx
@@ -17,11 +17,11 @@ import LikeMusic from '@/components/profileCard/likeMusic';
 import Post from '@/components/profileCard/post';
 import useProfileCardStore from '@/store/profileCardStore';
 import PostMessage from '@/components/profileCard/postMessage';
+import useGetRandomBackgrounds from '@/hooks/useGetRandomBackgrounds/useGetRandomBackgrounds';
 
 const MatchingPage = () => {
   // 배경색 목록
   const backgrounds = [
-    'bg-background-white',
     'bg-background-grey',
     'bg-background-yellow',
     'bg-background-green',
@@ -31,22 +31,7 @@ const MatchingPage = () => {
     'bg-background-sky',
   ];
 
-  // 이전 배경색 상태와 현재 배경색 상태
-  const [previousBackground, setPreviousBackground] = useState<string>('');
-  const [currentBackground, setCurrentBackground] = useState<string>('');
-
-  // 컴포넌트 마운트 시 또는 이전 배경색이 변경될 때 실행
-  useEffect(() => {
-    // 이전 배경색을 제외한 새로운 배경색 목록 생성
-    const newBackgrounds = backgrounds.filter((bg) => bg !== previousBackground);
-
-    // 새로운 배경색 목록에서 랜덤으로 하나 선택
-    const randomBackground = newBackgrounds[Math.floor(Math.random() * newBackgrounds.length)];
-
-    // 현재 배경색을 업데이트하고, 이전 배경색 상태도 업데이트
-    setCurrentBackground(randomBackground);
-    setPreviousBackground(currentBackground);
-  }, [previousBackground]);
+  const currentBackground = useGetRandomBackgrounds({backgrounds});
 
   // 잠금 상태
   const [isLock, setLock] = useState<boolean>(true);
