+++ src/type/customHook/useGetRandomBackgrounds.ts
@@ -1,0 +1,3 @@
+export type useGetRandomBackgroundsProps = {
+  backgrounds : string[];
+}
